add exports for node.js env
