+++ packages/elements/package.json
@@ -14,7 +14,8 @@
 	"types": "./index.d.ts",
 	"type": "module",
 	"exports": {
-		".": "./index.js"
+		".": "./index.js",
+		"./node": "./node.js"
 	},
 	"files": [
 		"*"
