+++ src/elements/attributes/shadow.ts
@@ -1,4 +1,4 @@
-import { isValidCustomElementName } from './tags.js';
+import { isValidCustomElementName } from './common.js';
 
 export const ShadowElements = [
 	'article',
