+++ src/elements/node.ts
@@ -1,8 +1,5 @@
-export * from './attributes/tags.js';
+
 export * from './attributes/common.js';
-export * from './attributes/attributes.js';
-export * from './attributes/shadow.js';
-export * from './attributes/form.js';
 
 export * from './ast/dom.js';
